streamlit/requirements.txt
pandas==2.1.4
streamlit==1.28.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pymongo==4.6.0
requests==2.31.0
minio==7.2.15